refactor(api): migrate competence-evaluation-serializer.js
